Rename class Controllers\Site\SiteController to Controllers\WebController
